requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
python-docx
openpyxl
fastapi
uvicorn
requests
pydantic